requirements.txt
aiohttp>=3.8.4
cryptography>=41.0.1
Cython>=0.29.35
easyocr>=1.7.0
getmac>=0.9.4
pandas>=2.0.2
Pillow>=9.5.0
psutil>=5.9.5
PyNaCl>=1.5.0
PyYAML>=6.0
selectolax>=0.3.14